[ignore] Modified module names by removing abs in aci_l4l7_service_graph_template_connection_to_connector and aci_l4l7_service_graph_template_connection
