+++ meta/runtime.yml
@@ -282,12 +282,17 @@ plugin_routing:
         removal_version: 3.0.0
         warning_text: Use aci_l4l7_service_graph_template_functional_connection instead.
     aci_l4l7_service_graph_template_abs_conn:
-      redirect: cisco.aci.aci_l4l7_service_graph_template_abs_connection
+      redirect: cisco.aci.aci_l4l7_service_graph_template_connection
       deprecation:
         removal_version: 3.0.0
-        warning_text: Use cisco.aci.aci_l4l7_service_graph_template_abs_connection instead.
+        warning_text: Use cisco.aci.aci_l4l7_service_graph_template_connection instead.
     aci_l4l7_device_selection_if_context:
       redirect: cisco.aci.aci_l4l7_device_selection_interface_context
       deprecation:
         removal_version: 3.0.0
         warning_text: Use cisco.aci.aci_l4l7_device_selection_interface_context instead.
+    aci_l4l7_service_graph_template_abs_connection_conns:
+      redirect: cisco.aci.aci_l4l7_service_graph_template_connection_to_connector
+      deprecation:
+        removal_version: 3.0.0
+        warning_text: Use cisco.aci.aci_l4l7_service_graph_template_connection_to_connector instead.
