+++ tests/integration/targets/aci_l4l7_service_graph_template_connection/tasks/main.yml
